prost_build: `BytesType and MapType` into a `collections` module.
